[refactor/post-demo-lite] rename /Models -> /businessLayer
