+++ android/ByteMe/app/src/main/java/businessLayer/RecordingFlag.kt
@@ -1,4 +1,4 @@
-package Model
+package businessLayer
 
 import android.support.annotation.ColorInt
